[Fix] tileset-tagger: fractional auto-fit fills the window (all tiles visible at any size) + map preview moved to a TALL left dock
Real cause of 'maximize hides tiles' + 'map render wrong':
- TilesetView used INTEGER zoom (min 1x): a 512px sheet in a ~700-1000px
  viewport was stuck at 1x -> tiny with empty space, and when a dock stole
  height it overflowed and cut off the bottom rows.  Switched to FRACTIONAL
  zoom so fitViewToWindow fills BOTH dimensions (shrinks below 1x when needed
  -> nothing is ever cut).
- A QScrollArea does NOT auto-resize a non-resizable child when its sizeHint
  changes; updateGeometry() was not enough, so the view stayed 320x320 and
  clipped the scaled sheet.  setZoom/setModel now resize(sizeHint()).
- The fit was read in resizeEvent BEFORE the central viewport was laid out
  (stale on maximize); now refit from an eventFilter on the viewport's resize.
- Map preview was a thin strip: tall/narrow route maps in a wide/short BOTTOM
  dock.  Moved Map-usage to a narrow+tall LEFT dock (like Tiled's map panel);
  fit-to-view shows the whole map readably.

+++ tools/tileset-tagger/TilesetView.hpp
@@ -1,7 +1,7 @@
 #ifndef TILESETTAGGER_TILESETVIEW_HPP
 #define TILESETTAGGER_TILESETVIEW_HPP
 
-// The tileset canvas: paints the .tsx image at an integer zoom with a tile grid,
+// The tileset canvas: paints the .tsx image at a (fractional) zoom with a tile grid,
 // tints tagged tiles by category colour, highlights untagged-non-empty tiles in
 // red (the owner's "reminder"), and lets the user rubber-band a rectangle of
 // tiles to tag.  Pure painting + mouse; the data lives in TagModel.
@@ -17,8 +17,8 @@ public:
     void setModel(TagModel *model);     // not owned
     void refresh();                     // repaint after the model changed
     void setShowStates(bool on);
-    void setZoom(int z);
-    int zoom() const;
+    void setZoom(double z);     // fractional so the sheet can fill the window AND shrink to fit
+    double zoom() const;
     void selectCell(int col,int row);   // programmatic selection (jump-to-untagged)
     QRect cellPixelRect(int col,int row) const;
     QSize sizeHint() const override;
@@ -38,7 +38,7 @@ protected:
 
 private:
     TagModel *model_;
-    int zoom_;
+    double zoom_;
     bool showStates_;
     bool selecting_;
     bool hasSelection_;
